+++ lib/ui/home/loading_photos_widget.dart
@@ -9,8 +9,8 @@ import 'package:photos/events/local_import_progress.dart';
 import 'package:photos/events/sync_status_update_event.dart';
 import "package:photos/generated/l10n.dart";
 import 'package:photos/services/local_sync_service.dart';
-import 'package:photos/ui/backup_folder_selection_page.dart';
 import 'package:photos/ui/common/bottom_shadow.dart';
+import 'package:photos/ui/settings/backup/backup_folder_selection_page.dart';
 import 'package:photos/utils/navigation_util.dart';
 
 class LoadingPhotosWidget extends StatefulWidget {
