+++ lib/ui/settings/backup/backup_section_widget.dart
@@ -8,13 +8,13 @@ import 'package:photos/services/deduplication_service.dart';
 import 'package:photos/services/sync_service.dart';
 import 'package:photos/services/update_service.dart';
 import 'package:photos/theme/ente_theme.dart';
-import 'package:photos/ui/backup_folder_selection_page.dart';
-import 'package:photos/ui/backup_settings_screen.dart';
-import 'package:photos/ui/components/captioned_text_widget.dart';
+import "package:photos/ui/components/captioned_text_widget.dart";
 import 'package:photos/ui/components/dialog_widget.dart';
 import 'package:photos/ui/components/expandable_menu_item_widget.dart';
 import 'package:photos/ui/components/menu_item_widget/menu_item_widget.dart';
 import 'package:photos/ui/components/models/button_type.dart';
+import 'package:photos/ui/settings/backup/backup_folder_selection_page.dart';
+import 'package:photos/ui/settings/backup/backup_settings_screen.dart';
 import 'package:photos/ui/settings/common_settings.dart';
 import 'package:photos/ui/tools/deduplicate_page.dart';
 import 'package:photos/ui/tools/free_space_page.dart';
