+++ lib/ui/settings/general_section_widget.dart
@@ -6,11 +6,11 @@ import 'package:photos/services/billing_service.dart';
 import "package:photos/services/feature_flag_service.dart";
 import 'package:photos/services/user_service.dart';
 import 'package:photos/theme/ente_theme.dart';
-import 'package:photos/ui/advanced_settings_screen.dart';
-import 'package:photos/ui/components/captioned_text_widget.dart';
-import 'package:photos/ui/components/expandable_menu_item_widget.dart';
+import "package:photos/ui/components/captioned_text_widget.dart";
+import "package:photos/ui/components/expandable_menu_item_widget.dart";
 import 'package:photos/ui/components/menu_item_widget/menu_item_widget.dart';
 import "package:photos/ui/growth/referral_screen.dart";
+import 'package:photos/ui/settings/advanced_settings_screen.dart';
 import 'package:photos/ui/settings/common_settings.dart';
 import "package:photos/ui/settings/language_picker.dart";
 import 'package:photos/utils/navigation_util.dart';
